+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [3.0.0] - 2021-08-05
+- Updated to `pronamic/wp-pay-core` version `3.0.0`.
+- Updated to `pronamic/wp-money` version `2.0.0`.
+- Switched to `pronamic/wp-coding-standards`.
+
 ## [2.1.1] - 2021-04-26
 - Happy 2021.
 
@@ -101,7 +106,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2014-12-12
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/2.1.1...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/3.0.0...HEAD
+[3.0.0]: https://github.com/wp-pay-gateways/sisow/compare/3.0.0...3.0.0
 [2.1.1]: https://github.com/wp-pay-gateways/sisow/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/wp-pay-gateways/sisow/compare/2.0.4...2.1.0
 [2.0.4]: https://github.com/wp-pay-gateways/sisow/compare/2.0.3...2.0.4